src/main/java/dao/jdbc/mapper/ProductRowMapper.java
package dao.jdbc.mapper;

import entity.Product;

import java.sql.ResultSet;
import java.sql.SQLException;
import java.time.LocalDate;

public class ProductRowMapper {
    public Product mapRow(ResultSet resultSet) throws SQLException {
        int id = resultSet.getInt("id");
        String name = resultSet.getString("name");
        int price = resultSet.getInt("price");
        LocalDate creationDate = LocalDate.parse(resultSet.getString("creation_date"));
        return Product.builder()
                .id(id)
                .name(name)
                .price(price)
                .creationDate(creationDate)
                .build();
    }
}
